+++ src/meltano/api/controllers/plugins.py
@@ -11,7 +11,6 @@ from meltano.core.project import Project
 from meltano.core.project_add_service import ProjectAddService
 from meltano.core.config_service import ConfigService
 from meltano.core.plugin_install_service import PluginInstallService
-from meltano.api.security import api_auth_required
 from flask_security import roles_required
 from meltano.api.api_blueprint import APIBlueprint