+++ src/meltano/api/security/__init__.py
@@ -7,13 +7,13 @@ from flask_security import Security, login_required
 from flask_security.utils import login_user
 from flask_principal import identity_loaded, Identity
 
-from .identity import users, FreeUser, ReadOnlyUser, create_dev_user
+from .identity import users, FreeUser, create_dev_user
 from .forms import MeltanoLoginForm, MeltanoRegisterFrom, MeltanoConfirmRegisterForm
 from .auth import (
     unauthorized_callback,
     _user_logged_in_hook,
     _identity_loaded_hook,
-    api_auth_required,
+    block_if_api_auth_required,
 )
 
 
@@ -29,14 +29,14 @@ def setup_security(app, project):
         "confirm_register_form": MeltanoConfirmRegisterForm,
     }
 
-    if not app.config["MELTANO_AUTHENTICATION"]:
-        if app.config["MELTANO_READONLY"]:
-            # the ReadOnlyUser is free to read, but not write
-            options["anonymous_user"] = ReadOnlyUser
-        else:
-            # the FreeUser is free to do everything and has all
-            # roles and permissions automatically.
-            options["anonymous_user"] = FreeUser
+    if not app.config["MELTANO_AUTHENTICATION"] and not app.config["MELTANO_READONLY"]:
+        # the FreeUser is free to do everything and has all
+        # roles and permissions automatically.
+        options["anonymous_user"] = FreeUser
+    else:
+        # Use Flask's built-in AnonymousUser which is not deemed to be authenticated
+        # and has no roles
+        pass
 
     security.init_app(app, users, **options)
     security.unauthorized_handler(unauthorized_callback)