+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 
 - [#1896](https://gitlab.com/meltano/meltano/-/issues/1896) Set pipeline update interval to daily by default, to start after first successful manual run
+- [#1888](https://gitlab.com/meltano/meltano/-/issues/1888) Explain in "Edit Connection" button tooltip why it may be disabled
 
 ### Fixes