+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 
 - [#1896](https://gitlab.com/meltano/meltano/-/issues/1896) Set pipeline update interval to daily by default, to start after first successful manual run
+- [#1892](https://gitlab.com/meltano/meltano/-/issues/1892) Clarify in Run Log modal that the "Explore" button can also be found on the Connections page
 
 ### Fixes